add BasicInboundXMLMessageListener
